Move displaydoc derive into integration test file
The Debug derive gets its own module to clearly demonstrate how the
snippets in the readme map to real code. The Display derive is just
there to test macro invocations; to the extent that it serves as
documentation, the relevant parts should make their way into doc
comments of the main reflect library.