Make references to generic parameters
This commit changes how generic parameters are stored. Instead of
storing a string value directly in a type or a function, a reference to
that paramter is stored there in stead. The actual parameter with the
original name is stored in a thread_local that can be accessed
throughout the crate. The reason for doing this is to be able to
differentiate between different generic parameters with the same name,
and also knowing when two parameters are actually the same. This will
become wery handy when doing some sort of type resolution, and
ultimately trait inference.