The content of a Function struct is only created once
When calling the SELF method for a function created by the library
macro, the content for that Function struct will be created only once
on the first call. A copy of that content is returned on any subsequent
call with that method. This is to ensure that the generic parameters
associated with that function is only set once in the TYPE_PARAMS and
LIFETIME variables located within the generics module.