Intruduce RuntimeTrait, RuntimeImpl and RuntimeParent
RuntimeTrait and RuntimeImpl are sub traits of RuntimeParent which
has a SELF method that returns an Rc<Parent>.

RuntimeTrait is used instead of RuntimeType in the make_trait_impl
method in Executions.

impl blocks for RuntimeType are no longer generated from the library
macro, since I couldn't come up with good way of dealing with generics
in type declarations. This shouldn't put restrictions on what one can
do with the library though, so it's probably fine.