Make fresh generic params for every function invoke
Whenever a function is invoked. A clone of that function with new
generic parameters is made for that invoke to differentiate between
the generics between every use of a funciton. This is because a generic
type can be bound to different types for every call, and this needs to
be kept track of.