Change how global data is handled
The way global data was handled previously was by having a global
WipFuncion that could be referenced when implementing a function. After
a function was implemented, the WipFunction was moved from the global
WIP variable to a WipImpl struct. Functions that took a reference to
anything that could be found in a WipFunction, had to rely on the
correct WipFunction being globally availible at the time that the
function was called.

This has not really been a problem up until I tried doing trait
inference. The trait_inference module needs access to the ValueNodes
of a function after the function has been completed, which means the
relevant WipFunction would no longer be globally availible. Calling any
function relying on that fact would crash the program. Not being able
to call certain functions at certain times, is not great, especially
when those functions are useful.

This commit unifies all the globally availible data to one place called
GLOBAL_DATA. It does make certian things a little less convenient,
since you need to know what data belongs to which function and impl,
and not having the data directly availible as was the case previosly in
the compiler module. The way it is handled now is by giving index
ranges to WipFunction and CompleteFunction so they know exactly what
data belongs to each function when they access data in the global data
store.